requirement.txt
requests
pytz
numpy 
pandas
yfinance
python-telegram-bot
nsetools

